@@ -1,7 +1,7 @@
 name: flipper_rw
 description: "A new Flutter project."
 publish_to: "none" # Remove this line if you wish to publish to pub.dev
-version: 1.170.4252223232688+1730365400
+version: 1.170.4252223232689+1730464370
 
 environment:
   sdk: ">=3.3.4 <4.0.0"
@@ -121,7 +121,7 @@ dependencies:
   firebase_ui_localizations: ^1.9.0
   flutter_localizations:
     sdk: flutter
-  flutter_localization: ^0.1.11
+  flutter_localization: ^0.2.2
   get_storage: ^2.1.1
   js: ^0.6.3
   flutter_web_plugins:
@@ -156,11 +156,11 @@ dependencies:
   flipper_services: ^1.0.0-dev
   flipper_login: ^0.0.1-dev.5
   # flipper_socials: 1.0.0+1
-  amplify_core: ^2.4.1
+  amplify_core: ^2.5.0
   amplify_api: ^2.4.1
-  amplify_flutter: ^2.0.0
-  amplify_auth_cognito: ^2.4.1
-  amplify_storage_s3: ^2.4.1
+  amplify_flutter: ^2.5.0
+  amplify_auth_cognito: ^2.5.0
+  amplify_storage_s3: ^2.5.0
   shared_preferences: ^2.3.2
   easy_localization: ^3.0.7 
   cloud_firestore: ^5.4.4
@@ -207,7 +207,7 @@ msix_config:
     force_update_from_any_version: true
   publisher_display_name: yegobox ltd
   identity_name: yegobox.yegoboxflipper
-  msix_version: 1.170.2836.0
+  msix_version: 1.170.2837.0
   publisher: CN=B0F334A1E-8AET-4B2C-BD1B-F239D4F791AE
   logo_path: "./assets/icon.png"
   start_menu_icon_path: "/assets/icon.png"
