@@ -46,15 +46,15 @@ dependencies:
   rxdart: ^0.27.7
   collection: ^1.18.0
   intl: ^0.19.0
-  amplify_core: ^2.4.1
+  amplify_core: ^2.5.0
   amplify_api: ^2.4.1
-  amplify_flutter: ^2.0.0
-  amplify_auth_cognito: ^2.4.1
-  amplify_storage_s3: ^2.4.1
+  amplify_flutter: ^2.5.0
+  amplify_auth_cognito: ^2.5.0
+  amplify_storage_s3: ^2.5.0
   async: ^2.11.0
   mime: ^1.0.6
   firebase_core: ^3.6.0
-  # powersync: ^1.8.2 
+  powersync: ^1.8.2
   permission_handler: ^11.3.1
   cloud_firestore_odm: ^1.0.0-dev.88
   json_annotation: ^4.9.0
@@ -68,6 +68,7 @@ dependency_overrides:
   http: ^1.2.2
   meta: ^1.12.0
   analyzer: ^6.0.0
+  
   
   
   web: ^1.0.0
